+++ extensions/done-bear/src/show-anytime.tsx
@@ -1,9 +1,8 @@
 import { withAccessToken } from "@raycast/utils";
+
 import TaskList from "./components/task-list";
 import { oauthService } from "./oauth";
 
-function ShowAnytime() {
-  return <TaskList view="anytime" />;
-}
+const ShowAnytime = () => <TaskList view="anytime" />;
 
 export default withAccessToken(oauthService)(ShowAnytime);