+++ extensions/done-bear/src/show-inbox.tsx
@@ -1,9 +1,8 @@
 import { withAccessToken } from "@raycast/utils";
+
 import TaskList from "./components/task-list";
 import { oauthService } from "./oauth";
 
-function ShowInbox() {
-  return <TaskList view="inbox" />;
-}
+const ShowInbox = () => <TaskList view="inbox" />;
 
 export default withAccessToken(oauthService)(ShowInbox);