+++ extensions/parachord/src/open-parachord.ts
@@ -1,0 +1,5 @@
+import { openParachord } from "./utils";
+
+export default async function Command() {
+  await openParachord("home", [], {}, "Opening Parachord");
+}