+++ extensions/bitwarden/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Update bundled Bitwarden CLI from v2025.11.0 to v2026.2.0 to fix "Invalid session token" error caused by server-side KDF upgrades
 - Fix incorrect vault status saved after API key login
+## [Fix] - 2026-04-02
+
+- Catch Invalid session token error
 
 ## [Fix & Improvements] - 2026-03-24