+++ extensions/netbird/CHANGELOG.md
@@ -1,5 +1,20 @@
 # NetBird Changelog
 
+## [Networks management] - 2026-04-02
+
+- Added `Show Networks` command to list network routes and select/deselect them from Raycast.
+- Fixed error which failed update command during asking for administrator permissions.
+
+## [Status on empty dns] - 2026-03-30
+
+- Fixed status command error when user does not have netbird DNS enabled
+
+## [UX Improvements] - 2026-03-27
+
+- Status screen now shows empty screen during loading instead of empty data
+- Homebrew errors handling and display during update
+- Automatically restart NetBird service with administrator privileges after a successful Homebrew update
+
 ## [Initial Version] - 2026-03-25
 
 - Initial release of the NetBird Raycast Extension.